NODE-1049 API method to get validation time of the script
