+++ src/main/scala/com/wavesplatform/Application.scala
@@ -245,6 +245,7 @@ class Application(val actorSystem: ActorSystem, val settings: WavesSettings, con
                         settings.blockchainSettings.functionalitySettings),
         DebugApiRoute(
           settings,
+          blockchainUpdater,
           wallet,
           blockchainUpdater,
           peerDatabase,
