+++ src/test/scala/com/wavesplatform/http/DebugApiRouteSpec.scala
@@ -9,7 +9,7 @@ class DebugApiRouteSpec extends RouteSpec("/debug") with RestAPISettingsHelper w
   private val wavesSettings = WavesSettings.fromConfig(sampleConfig)
   private val configObject  = sampleConfig.root()
   private val route =
-    DebugApiRoute(wavesSettings, null, null, null, null, null, null, null, null, null, null, null, null, configObject).route
+    DebugApiRoute(wavesSettings, null, null, null, null, null, null, null, null, null, null, null, null, null, configObject).route
 
   routePath("/configInfo") - {
     "requires api-key header" in {
